Provide a simple Vagrantfile
This Vagrantfile:

* Refers to Ubuntu 14.04, which has a recent (3.13) kernel.
* Uses install.sh to provision the box.

This does not forward any ports other than 6080, and although
it does make the box appear on a private network with the host,
since the install script by default binds all services to localhost,
it is not fully usable for app development.

I tested that 'vagrant up' works.