Allow running the Sandstorm server as non-root.
If the bundle is run as a non-root user, the bundle runner automatically switches to using UID namespaces. Unfortunately, there is currently no way to make "spk dev" work in this mode, as FUSE filesystems currently cannot be created inside UID namespaces.

The auto-updater is automatically disabled if the install location cannot be written by the running user, but works fine as long as the user owns everything.

The installer script has NOT yet been updated to offer non-root installs.