Update to Meteor 0.8.2 and massively reduce build time.
Meteor 0.8.2 includes my patch that allows "meteor bundle" to output a directory rather than a tarball, avoiding a compression step that's a waste of time for us since we just decompress the bundle immediately anyway.

I also added the XZ_FLAGS make variable which when set to -0 cuts about 40 seconds (i.e. most of it) off the build time.