Fix the random '403 Unauthorized' problem that often happened when creating a grain right after deleting one.
Turns out the problem is that if the new grain gets the same port as the old, the browser might reuse an existing connection it made before, which actually leads it to the old, thought-to-be-dead proxy object, which promptly fails. Solution is to wait for all connections to be closed before reusing a port. Would be better if Node provided a way to force-disconnect them, though.