Auto-assign a subdomain for web publishing apps.
This is mainly to improve the demo experience, where you probably don't want to play with your DNS settings just to see what the site will look like. Moreover, our demo server doesn't currently support mapping arbitrary domains to it, so this patch lets the app discover whether the account is a demo account so that it can display a note that domain mapping won't work, but the auto-assigned host will.