Make installer default to BASE_URL local.sandstorm.io instead of localhost.
This is important because the WILDCARD_URL will have to be *.local.sandstorm.io since there's no way to set a wildcard in /etc/hosts (which is how 'localhost' is defined). If these two URLs aren't under the same domain, browsers or extensions that block "third-party cookies" will break Sandstorm.

Fixes #78.