Fix stuck-in-analyzing on app install.
Our database writes were being done as fire-and-forget, without waiting for completion before continuing. The "analyze" step, meanwhile, can complete very fast -- in fact, it doesn't usually block at all. This means the transition to "analyzing" and the subsequent transition from "analyzing" to "ready" would occur so close together that they could end up received in the wrong order by Mongo. Fix is to wait for the previous call to complete before issuing a new one.

Fixes #70. (At least, I hope so. It's hard to know for sure since the problem is rare.)