Transparent IP proxying: Setup supervisor to route outgoing packes back to the sandbox.
After this change, any attempt inside the sandbox to connect or send datagrams to a non-localhost address will be routed to 127.0.0.1:23136. Some process inside the sandbox can listen on that port and act as an IP proxy. E.g. such a process could request an IP capability via the Powerbox and proxy through that. We plan to extend sandstorm-http-bridge with this functionality.

The ip_tables kernel module must be loaded for this to work. If it's not loaded, the supervisor simply will not enable the routing. We should extend run-bundle and/or the Sandstorm front-end to detect this problem, as well as the Sandstorm installer.