In the sandbox, bring up a dummy interface and route packets through it.
Our previous attempt to route packets through "lo" (and then use iptables to capture them and redirect them to a local port) did not *quite* work: the packets were captured, but reply packets did not seem to get back to the sender. Using a "real" (i.e. non-loopback) interface seems to fix this. We create an interface of type "dummy".