Fold minibox into main sandstorm binary.
This leaves only sandstorm-http-bridge as a separate binary, which makes sense since http-bridge is only included to be bundled into apps.

Before building a new release, we need to enable static linking for both of these binaries. Unfortunately that's not quite possible with Ekam at the moment; will need to fix.