Create /appdemo/ to launch a demo instance of an app
The URL structure is /appdemo/:packageId (which is the sha256sum of the
package).

For now, we ignore any query string data provided. Crucially, we ignore a
?url= parameter, which /install/ uses -- since the app demo only works for
apps that are already installed, need no ?url parameter.

This also extracts the code to launch and enter a grain, given either a
package ID or an action ID, into a helper function.

Close #211