Hack fix #224.
An arguably better fix might be to install capnp/*.capnp into node_modules as part of the ekam build, but it's a bit weird to put such a rule into Cap'n Proto's code since it is not Node-specific, and it's also weird to put it in Sandstorm's code since it's Cap'n Proto's files. Punt for now.