Add ApiSession
Now when calling newSession from a proxy that was initiated from an API
connection, we now call it with a type id of ApiSession.

As a transitional measure, we fallback to WebSession if the call asking
for an ApiSession fails. This will eventually be removed once all apps
have updated.

Added apiPath as a variable to BridgeConfig. This will enable ApiSession
functionality as well as limiting what paths an api token can access.