Add email token login system
This is an alternate login system, based loosely on accounts-password.

Users input their email, and the server sends a message to that address
containing a token. The user either clicks a link out of that email or
manually copies the token into an input box, and then they are logged
in. This package (accounts-email-token), was intended to be generic
enough to be used outside Sandstorm, and may at some point be moved out
and put onto atmosphere.

On the Sandstorm side, the only things that needed to chage were the
admin page, and the accounts-ui. It's a bit of a mess since we're
building on top of meteor's accounts-ui that has a bunch of custom logic
for accounts-password that I tried to re-use as much as possible.