Add token templating
Grains can now make a request to the shell via postMessage() with a template
which can substitute a privileged API token for $API_TOKEN and the API host for
$API_HOST.  The grain will receive a postMessage reply from the shell with a
single-use URL in the shell's origin that will display the rendered template.
The app should iframe this URL in a place in the document inline where it wants
the template to appear.  Apps should *not* persist this URL anywhere.

Implementation detail: the privileged token will expire 5 minutes after
creation, if it is not used in that timeframe.  This helps avoid cluttering the
DB (and UI) with a bunch of tokens that were generated for this templating magic
that the user didn't end up using.