Refactor ongoing notification logic in supervisor
Drop the usage of random wakelock ids, and now only store wakelocks in a
table when `save` is called on one. The dropping of the wakelock is now
leveraged on the object graph and relying on the fact that Capabilities
are automatically ref counted.

The only complication added to accommodate this is a weakref between the
handle and the wrapped ongoing notification. This is needed in order to
not rely on the frontend actually ever calling cancel or dropping the
notification in the case of the handle being dropped.