Emergency fix: Make enabled services reactive to solve race condition.
Currently, no one can log into alpha because the login drop-down is disabled if no services are enabled, the subscription which indicates that they are enabled arrives after the template initially renders, and the list is not reactive. Therefore, everyone sees "no login services are configured".

This apparently wasn't noticed in local testing because when the server is local, the subscription often comes back before the login drop-down first renders, and everything appears to work.

This change also gets rid of pointless differences in behavior when only one service is enabled, and when no services are enabled the user is offered a link to the admin page to fix it.