Add initial docs for Sandstorm + vagrant-spk
The primary purpose of this commit is to store a vagrant-spk
packaging tutorial somewhere that people can find it. It also creates
a structure for future Sandstorm documentation, organized similarly
to the Django documentation.

If you have mkdocs installed, you can generate HTML based on this
by being in the Sandstorm directory and running:

mkdocs