Add two nudge arrows for demo users
For demo users with no grains, point an arrow at "Install apps" and
"New App Instance" if they are looking at an app.

This visually conflicts with the information about the managed hosting
and mailing list signup, so we hide that information until the demo user
has created a grain.