Fix bug where all shared grains would force-refresh on server update.
The problem is that added() is called for everything on startup.

It's not necessary for security to refresh on added(). It is slightly sad that users have to manually refresh when they are granted new permissions, but force-refreshing them could cause them to lose work, so we really need a third option.