Add opt-in passthrough of remote address to ApiSession as X-Real-IP
While in most cases, we want to avoid exposing a client's IP address to
sandboxed apps, having the option of sharing it is acutely valuable for
analytics apps like Piwik, which benefit substantially from having external IP
addresses for geolocation.

If an authenticated client sends the API endpoint a header:

    X-Sandstorm-Passthrough: address

then we will pass the client's observed IP address as a parameter in the
ApiSession, and sandstorm-http-bridge will pass that to the app's backend as
an X-Real-IP header.

Additionally, if the request to the API endpoint is made from an RFC1918
address, we will substitute the value of an X-Real-IP header, if present, for
the client's address.  This should make things work out-of-the-box for most
simple SSL termination setups, while preventing spoofing of the address seen by
apps by anyone on a separate network from the sandstorm server.

We only allow this for API sessions because doing so for web sessions would
allow a malicious app to serve JS that makes requests back to the app with the
opt-in header, leaking the user's real IP without first requiring user action.
This would break our sandboxing guarantees.