Implement web publishing by RPC rather than filesystem access.
I've added a hacky method to the Supervisor interface, so that the supervisor can directly serve these files. I also modified the supervisor to delay starting up the app until it receives a getMainView() or restore() call, which requires the app to be running. This allows WWW files to be served without actually starting the app. The overhead of running the supervisor is comparatively small, and is acceptable for the time being until we replace all of this with a driver.

I believe this eliminates the last way that the front-end was reaching directly into grain storage.

This change will make web publishing work correctly on Oasis.

All code in this change will go away as soon as we have web publishing drivers. There are many improvements that could be made otherwise, e.g. caching and range requests, but they aren't worth the effort for throw-away code.