Fix Makefile rule for update-deps.
This never worked right as-is because DEP was a shell var while REMOTE_* were make vars, so $(REMOTE_$(DEP)) was evaluated before the for loop ever ran and always evaluated to an empty string. This appeared to work because it ended up using the default origin for each repo. Fixed by using a gmake `foreach` rather than a shell `for`.