App index with `spk publish` command.
The app index is itself a Sandstorm app. It presents an API through which apps can be submitted (with the new `spk publish` being the client to this API) and presents a review interface to the admin. Once apps are approved, they are published via web publishing such that the app market client can download and display the index, and Sandstorm servers can install packages from teh index.

Currently `spk publish` only works with a webkey pointing to an app index instance, but once we have a standard one up we'll embed the webkey directly into the spk source code.