Fix billing prompt scrolling.
The inner scroll div simply didn't work. There's no one set of styles that makes scrolling work when the window is small without leaving ugly empty space when it is large.

So we give up and put the scroll bar on .frame. This means that the trusted UI caret disappears, because it's outside the overflow. We should fix that later by making the carets be siblings of .frame instead of children.