Redirect home to new grain page, or search grains, depending...
If you are an invited user or a demo user, and you have no grains yet,
you go to new.

Otherwise, you go to search.

Note that this moves the subscription to grainsMenu to be global, to avoid
races. It therefore removes subscribing to that from a couple of places.