Move the app submission instructions into docs
I am not sure where to place this, so far this is just the basics, and it's not linked to other pages yet. I've combined details from the Github wiki and the package.capnp file to create this, and it's possibly worded in a friendlier manner? Licenses should be broken out into a separate page, almost certainly. This still links to the Etherpad file as well as package.capnp as it is not as exhaustive, nor does it currently show formatting for the manifest. (I highly recommend spk init creates manifests with metadata fields ready to go. (See also: #787)