Avoid duplicate subscriptions, which it turns out are not deduped by meteor.
Meteor de-dupes the *results* of subscriptions -- i.e. avoids sending down documents that it knows the client already has -- but that happens only after it has repeated the mongo query on the server and checked each document to see if the client already has it. So, we were created a lot of redundant server load with these dupe subscriptions.