Update links on Raw Packaging Guide
A lot of links led to the Github wiki, which then redirected you back to the docs. Also, remove "Install Vagrant" option for raw packaging, because those users should be using vagrant-spk, and we introduce people to that first, including directing people to it at the top of this page.