Cryptographically verify updates to Sandstorm proper.
Updates are signed using ed25519 via libsodium. We don't use gpg for updates because it would be an extra dependency and it's hard to use correctly. (OTOH we have to use gpg for install verification since we need to bootstrap from tools people have; see pull request #901.)

The strategy is as follows:
- We maintain a list of signing keys, which for now has only one element.
- The public keys in this list are compiled directly into the Sandstorm binary.
- Private keys are kept in a keyring file.
- Every time we want to "rotate" the key, we actually add a new key to the list, and use it to sign future updates.
- Each update is shipped with a signature file containing signatures using *every* key in the list.
- Sandstorm, when updating, checks the signature file for a signature using the most-recent key it knows about.

The scheme makes it painless to rotate keys, since rotating keys will never prevent old versions from updating.

In the event of a compromise of the update signing key, the update system is reduced to relying on HTTPS for integrity until a new release is pushed with a new key added -- which would be almost a non-event, honestly.

In the event that the private signing key becomes inaccessible to us, it will become impossible to update existing clients without manual user intervention -- which would be catastrophic.

Therefore, we should err on the side of keeping many copies of the private key.