Document PGP-verified install.
This is (not entirely surprisingly) overcomplicated. It could perhaps be less complicated if we could find a more idiomatic way to "certify" release keys. However, GPG does not seem to be designed to handle quickly-expiring, frequently-renewed certificates:
- Its UI does not want to let us issue a new certificate while an existing one is still valid.
- When verifying, an expired certificate is still displayed with only a cryptic extra 'X' character to indicate that it is expired.

For now, we are falling back to natural-language certificates instead, but this means that the user must manually verify the key fingerprint, which is itself surprisingly hard through the GPG UI, assuming we don't want to suggest that users use --status-fd and try to manually interpret the output.