Move fallback logic about packages to SandstormDb static methods
We have a decent amount of logic dealing with backwards-compatibility with app
titles and actions, and guessing values for new manifest fields based on older
ones.  This commit attempts to centralize that logic as some methods on the data
model in the sandstorm-db package.

This commit also add an appId field to devApps, to make accessing them more
like accessing Packages.  This makes dealing with packages that may be dev
packages or installed packages a bit simpler.