Merge GET and HEAD into a single method.
We use a boolean parameter to determine whether the caller wants the entity body.

This makes life easier for people trying to implement the raw API -- they don't necessarily need to pay attention to this parameter, and the front-end will drop the content for them.

It also means that existing spks will properly support HEAD without changes.