Be more specific about WebDAV method inputs.
Many of these methods take some XML as the input, so we might as well pass that as a string. (In theory we should actually represent the XML structure but no way am I going to add an XML parser to our dependencies).

The UNLOCK method appears to specifically take a "lock token" as input (in fact it appears before this change the token was being lost).

For the couple of methods that take arbitrary content, I'm reusing PostContent. When we do an API revision, we should replace PostContent and PutContent with a single Content struct, rather than repeat ourselves.