Copyediting of "Sandcats dynamic DNS"
"...which Sandstorm (and the install script) manage for users" - subject and predicate should agree in the absence of the parenthetical; here, they do not. See [The Chicago Manual of Style](http://www.chicagomanualofstyle.org/qanda/data/faq/topics/Usage/faq0180.html) for more details.

"...and places it in places it in" - typo.

"To manually recover..." - use spacing consistent with previous subsection.

Add mailto link for the lazy.