Install flow shows app details
Adapted the app details components to allow for reuse in the install flow.

If "safe", installing an app proceeds directly to the app page after successful
download and installation.  If not known safe, users are prompted with the details
of the app before confirming installation.

Fixes #15 (as of this writing, the oldest open bug!)
Fixes #1043
Fixes #1135 as part of the refactor
Closes #873