Picker powerbox for SealedUiView
Apps can now specify queries with their powerboxRequest `postMessage()`s.
If they specify that they want a SealedUiView, then we have some UI to let
the user select a grain and a RoleAssignment, then we `postMessage()` back
a SturdyRef token restorable to a SealedUiView.

Restoring the SealedUiView into a liveref and then offering that liveref on a
user's session context will share that grain with them, with the specified
roleAssignment.

The picker powerbox classes and templates are designed to make it fairly easy
to plug in additional custom UI for various other interfaces which will gain
frontendRef providers.