Add a C.UTF-8 locale descriptor, and generate a locale in make-bundle.sh
Before this patch, to obtain a sane locale, we would copy in locales from the
host system, then set `LANG=C.UTF-8` in `run-bundle.c++`.  This had two issues:

* Build hosts that aren't running Debian don't have a C.UTF-8 locale, so the
  generated bundle wouldn't run unless you'd manually defined one on the host
* This included *all locales* from the host system, including the whole
  locale-archive, which (if you have additional locales installed) could add
  ~100MB to the package.

With this patch, we move to standardize what locale data we ship:
`make-bundle.sh` will generate a locale similar to Debian's C.UTF-8 locale for
inclusion in the bundle.  This locale, like Debian's, is non-conforming to the
SUS, but is significantly more useful in a variety of ways that are more
important for real applications, such as:

* having actual correct character classes for the whole Unicode character set
* having a guaranteed sort order for characters beyond U+007F

To minimize duplicated effort delta from glibc upstream, this implementation
still relies on the system's UTF-8 character map and some other glibc-provided
i18n files that were available as part of the minimal install on all systems I
tested.

Closes #169.