Add a test for grain backup and restore
Dealing with downloads and uploads in selenium is...fraught.

You can't interact with any of the native platform dialogs, so you have to
configure Firefox to auto-download files of a particular MIME type to a fixed
location via a profile which you also have to configure on a per-test basis.

You can't really know when a download is completed, since there's no
programmatic notification of completion available.  We watch the hardcoded
download folder for changes, and assume that since our backup is small, it will
complete within 5 seconds of clicking the button, and within 1 second of the
file being initially open()'ed.

You can't set the value of a <input type="file"> element in Firefox if the
element isn't visible.  The shell makes them invisible by default and supports
interaction by triggering input.click() when you click a parent button.  So to
make this work in testing, we patch the style of the <input>.

Finally, we need to refactor the shell to allow the tests to call the
appropriate methods to trigger form submission, since we normally activate the
<input> and wait for its value to change (but we can't here, because Selenium
can't interact with the native platform UI like the file-picker).

A new test app ( https://github.com/zarvox/sandstorm-test-python ) was written
to support this test and others.