On initial signup, if billing is enabled, show a billing plan selector.
This lets the user know that billing is a thing. The user can of course choose "free", but perhaps some will want to choose a plan immediately.

The strategy is similar to the profile confirmation page, in that the plan selector overrides other routes. However, there is an exception: if the user is trying to view a shared grain (not one they own), then we do not interrupt them, since this would be weird for people who are only here to interact with their friend's grains.

There is a risk that this change will scare some users away too early. We should consider running a controlled experiment.

This is the Sandstorm side of the change. Most of the meat is in Blackrock.