Enable specifying a background for the logged-out home page.
If specified, the URL will be displayed in an iframe behind the prompt. This allows every server to customize their home page.

I removed the old "splashDialog" setting as it had become kind of ugly and now seems unnecessary. Plus, it technically allowed the admin to XSS users. While the legitimate admin almost certainly has this power anyway, limiting the power of the admin panel makes hijackings less damaging.