Remove "menuitem" role from login buttons
menu/menuitem roles are a bit picky in how they are used. If non-menuitem children appear in menus, screen readers don't know how to present that and generally make a bad job of it. In this instance you're displaying the email form as a child of the menu, which a) isn't a menuitem and b) doesn't present well if you make it a child of a menu.