Modernize .meteor/packages
standard-app-packages was renamed to meteor-platform in August.
meteor-platform was deprecated with the release of Meteor 1.2.
Per the upgrade guide ( https://quip.com/RXFlAk9Rc2xI ), we
should explicitly name the packages we wish to depend on.

I've taken this opportunity to also add the ecmascript and
ecmascript-collections packages, which will allow us to start using a number of
nice ES2015 syntax features - I'm most excited about arrow functions and
classes.