Add a pause to roundcube test to increase reliability
Roundcube appears to discard an early click on the settings button.  The click
changes the button's appearance, but does not actually lead to the settings
pane being shown, which makes the rest of the test fail.

Adding a short pause to allow the initial pageload to complete seems to make this
better.