Make popup .frame be a flexbox.  Make the close button not a float.
Picker powerbox needs .frame to be a flexbox so it can automatically grow where
appropriate.

I'd have stuffed that logic inside a powerbox-specific inner div, but then (for
reasons I do not understand) the `max-size` on .frame to keep it from covering
the whole UI doesn't limit the size of the inner div.