`spk publish` needs to use the new key to sign requests, and the app index needs to expect this.
This includes requests that modify the status of old packages originally signed with the old key.

For that matter, it would be nice if the app ID as listed in sandstorm-pkgdef.capnp can continue to be the old app ID, and spk automatically uses the new key as appropriate. Hence, this modifies `spk pack` as well.