Memoize results of newApiToken for offer templates for a short time.
This ensures that when Davros uses multiple offer templates in the same page, they use the same token (and therefore host).

It's also nice that this means we won't generate unnecessary tokens when someone browses away from the offer template page and back repeatedly.

I added a client-side copy of the es6-promise polyfill because promises make this code much cleaner, and our existing copy of the polyfill is server-only and will go away as soon as we upgrade to node4. Also, weirdly, the existing Meteor `promise` package is server-only.

I tested on IE11 to verify that the promise polyfill works on expected (all other browsers support Promises natively).