Adjust location and content of experimental index.
This makes the experimental index contain only pending apps. Rejected apps will not be included.

The hack in Sandstorm to enable CORS on the index was not handling the experimental index properly, so that is fixed here too, but a new Sandstorm push will be needed before it can be used.