Restore last-opened grains on load.
We store the list of open grains in localStorage at shutdown and restore it at startup. We also clear the saved list at startup so that if you open a second window it doesn't open the same grains again. Additionally, we do *not* store anything when closing a window that only has one grain open (or none), as probably in this case the user is relying on browser tabs rather than our sidebar tabs and restoring the last-opened grain would be surprising.

(Also fixed a random exception I've been seeing and fixes #1578.)