Refactor Picker Powerbox so that query matching happens on the server side.
PowerboxDescriptors are now provided in base64url-serialized-packed Cap'n Proto format, not JSON.

This will be necessary going forward because PowerboxDescriptor.Tag.value is an AnyPointer, the schema of which Sandstorm may not even know. The query-matching algorithm is designed to operate on Cap'n Proto binary format without the need for a schema, but without a schema it is not possible to convert to/from JSON.

I expect most apps won't need to customize their queries very much, so can probably get away with using hard-coded queries copied from documentation. Copying a whole query string seems not much worse than copying a type ID.