Fix mobile popup positioning
We rely on `position: static` on mobile to ignore the `left: $41px` or `right:
0px` inline styles which we compute to place the popup near its controlling
element for desktop.

Since we benefit from using `position: relative` for the contents of the frame,
split this into two nested divs:

`div.frame-container` is responsible for the placement and size of the popup as
well as border and drop-shadow.

`div.frame` is responsible for including all of the interesting content of that
popup, which can be `position: relative` to allow for absolutely-positioned
items within the frame.

Fixes #1725.