Setup wizard
To explain some of the unusual code and layout: this changeset started out as
an attempt to build the new admin UI, and switched to being the setup wizard
partway through.  A significant amount of code for the former was reusable for
the latter; rather than try to separate out the two changes, I include some
unrelated unlinked routes (`/admin-new/*`) and supporting code
(`shell/client/admin/*`).

We're going to need that code in the not-so-distant future anyway, and I didn't
really want to deal with figuring out exactly what depends on what, pulling the
pieces apart, only to merge them back together later.