Offer templates should avoid extra newlines on the clipboard in Firefox
There appears to be a Selection edge case in Firefox that inserts spurious
newlines into the value copied onto the clipboard if you attempt to select
everything contained by a <pre> within a <div> if the <div> contains newlines,
but the <pre> does not.

Removing the contained newlines makes Firefox do the right thing.

Fixes #1795.