Move ApiTokenOwner.user.lastUsed onto ApiTokens proper
It seems likely we'll be interested in tracking when *any* kind of ApiToken
gets used, so we can show more useful information in the connections UI.
Additionally, it doesn't really make sense to be passing around lastUsed in
calls to save() or other things that take an ApiTokenOwner.  So move the field
out of token.owner.user and onto token itself.