Improve grain renaming behavior. Fixes #1087
Historically, when a grain is shared, the title would be copied and become the receiver's "petname" for the grain, displayed whenever the receiver opens the grain. If the owner renamed the grain post-sharing, the receiver would never see the rename, only their petname. If the receiver renamed the grain, it would change their petname. This was generally confusing to people.

However, the alternative of "always show the owner's title" is also problematic since:
- If the title changes in the recipient's grain list, they may be confused as to how this new grain that they don't remember got here.
- If the owner's title is poorly-chosen or similar to other grains in the receiver's grain list, the receiver may want to retitle the grain for themselves (hence the petname behavior).

This change implements a compromise by displaying both the petname and the owner's title:
- If the petname and original title are the same, we just display it.
- If the sharing recipient has explicitly modified their petname, we show their petname as the title, with subtext of "renamed from: [upstream title]". This allows the user to assign their own title if they need to disambiguate the grain, though most of the time they probably won't.
- If the sharing recipient has not explicitly modified their petname (i.e. it is a clone of the title from the time of sharing), but the upstream title has changed, we show the upstream title, with subtext of "was: [petname]". This allows the user to understand how this grain whose title they don't recognize arrived in their grain list.