Refuse to call meteor build if broken symlinks are present.
Meteor will silently ignore broken symlinks, as if they weren't there. This can lead to introjs.css being missing from the build, particularly when building Blackrock which itself sets up the "shell" directory as a symlink forrest back into Sandstorm's "shell" directory. node_modules may be present in the Blackrock directory, allowing meteor to find and build dependencies, but may be missing in the Sandstorm directory, causing the introjs.css symlink to be broken, causing that CSS to be silently dropped.

Fixes #1899